pvr: correct the date type [1/1]
PD#OTT-14805
PD#SWPL-37607

Problem:
the data type is incorrect, which caused overflow.

Solution:
correct the date type

Verify:
franklin

+++ am_adp/am_tfile/am_tfile.c
@@ -1211,7 +1211,7 @@ loff_t AM_TFile_Tell(AM_TFile_t tfile)
 
 loff_t AM_TFile_GetAvailable(AM_TFile_t tfile)
 {
-	int ret = 0;
+	loff_t ret = 0;
 	if (! tfile->opened)
 	{
 		AM_DEBUG(0, "[tfile] has not opened");
